refactor: snake_case named files and fixing imports
