+++ lib/pages/boarding/second_boarding.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import '../../utils/styles/colors.dart';
-import '../../widgets/AppLargeText.dart';
-import '../../widgets/AppScreenTitle.dart';
+import '../../widgets/app_large_text.dart';
+import '../../widgets/app_screen_title.dart';
 
 class SecondBoardingScreen extends StatelessWidget {
   const SecondBoardingScreen({super.key});
